requirements.txt
discord.py>=2.3.2
py-cord>=2.4.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
httpx>=0.25.1
PyJWT>=2.8.0
python-multipart>=0.0.6
cryptography>=41.0.7
aiosqlite
cryptography
discord.py
fastapi
httpx
py-cord
PyJWT
python-dotenv
python-multipart
uvicorn[standard]